+++ StudyTimer/Views/TimerCountdownView.swift
@@ -6,6 +6,7 @@
 //
 
 import SwiftUI
+import Combine
 
 struct TimerCountdownView: View {
     @ObservedObject var viewModel: ViewModel
@@ -21,11 +22,12 @@ struct TimerCountdownView: View {
             Circle()
                 .trim(from: 0.0, to: viewModel.progress)
                 .stroke(style: StrokeStyle(lineWidth: 20, lineCap: .round))
-                .foregroundColor(.cardBackground)
+                .foregroundColor(viewModel.progressColor)
                 .rotationEffect(.degrees(-90))
             
             Text(viewModel.timeString)
                 .font(.system(size: 30, weight: .semibold, design: .monospaced))
+                .foregroundColor(viewModel.progressColor)
         }
         .frame(width: 150, height: 150)
         .onReceive(timer) { _ in
@@ -36,32 +38,65 @@ struct TimerCountdownView: View {
 
 extension TimerCountdownView {
     final class ViewModel: ObservableObject {
-        @Published var remainingTime: Int
-        @Published var isRunning: Bool = false
+        internal var isRunning: Bool = false
         let totalTime: Int
+        @Published var remainingTime: Int
+        var completedSubject: PassthroughSubject<Void, Never> = .init()
 
-        init(totalTime: Int,
-             isRunning: Bool = false
-        ){
+        init(totalTime: Int){
             self.totalTime = totalTime
             self.remainingTime = totalTime
-            self.isRunning = isRunning
         }
-
-        func tick() {
-            if isRunning && remainingTime > 0 {
-                remainingTime -= 1
-            }
-        }
-
+        
         var progress: CGFloat {
             CGFloat(totalTime - remainingTime) / CGFloat(totalTime)
         }
-
+        
         var timeString: String {
             let minutes = remainingTime / 60
             let seconds = remainingTime % 60
             return String(format: "%02d:%02d", minutes, seconds)
         }
+        
+        var progressColor: Color {
+            var opacityValue: CGFloat = 0.0
+            if progress <= 0.25 {
+                opacityValue = 0.25
+            } else if progress <= 0.5 {
+                opacityValue = 0.5
+            } else if progress <= 0.75 {
+                opacityValue = 0.75
+            } else {
+                opacityValue = 1.0
+            }
+            return .cardBackground.opacity(opacityValue)
+        }
+
+        
+        func start() {
+            isRunning = true
+        }
+        
+        func stop() {
+            isRunning = false
+        }
+        
+        func reset() {
+            remainingTime = totalTime
+        }
+        
+        func tick() {
+            guard isRunning,
+                  remainingTime > 0
+            else {
+                return
+            }
+            remainingTime -= 1
+            if remainingTime == 0 {
+                isRunning = false
+                completedSubject.send(())
+            }
+        }
+
     }
 }
